FIX: Dependencies: FLANN pkgconfig directories
Currently, pkgconfig only looks in */lib64/* directories, causing FLANN to not find lz4 on Apple targets. This might work on Linux, but on Apple platforms the differentiation between lib64 and lib does not existby default. This commit allows pkgconfig to look in */lib/* as well, as this is the default installation directory on macOS. The path is prepended, so on Linux targets */lib64/* should still take precedence.

Signed-off-by: Philipp Remy <philipp.remy@law-school.de>