FEAT: Dependencies: Allow building an embedded XercesC library
XercesC is a required hard dependency of libE57Format and it cannot build as an embedded project. Allow the user to rely on building a XercesC library as a dependency. Because it is *only* required by libE57Format, make it a cmake_dependent_option() based on whether libE57Format should be build.

Signed-off-by: Philipp Remy <philipp.remy@law-school.de>