FIX: AliceVision configure variables
Explicitly pass through top-level CMake variables when building AliceVision as an external project in Dependencies.cmake.

This change allows for the following configure step to respect any choices the user made at the top level.

FIXME: We should ideally only have one place to set the options - the very top-level CMakeLists.txt. But that would be a breaking change, as we would need to remove all "AV_" prefixed variables and replace them with the src/CMakeLists.txt "ALICEVISION_" prefix style.

Signed-off-by: Philipp Remy <philipp.remy@law-school.de>