FIX: Use common rpaths on Apple targets
Adapt the install rpaths on Apple targets to accommodate for the typical library paths.
This includes the standard Unix paths (lib, next-to-loading-binary) and the Apple bundle structures (Libraries, Frameworks). Add every possibility for @loader_path (for dynamic libraries) and @executable_path (for executables) to ensure proper lookup based on the context.

Signed-off-by: Philipp Remy <philipp.remy@law-school.de>