FIX: Dependencies: Bump ffmpeg version
Bump ffmpeg version from v5.1.2 to v7.1.1. This should be fine, as OpenCV and OpenImageIO appear to be the only dependants of ffmpeg, with both having support for ffmpeg 7.X.

This fixes an issue on x86_64 Linux where inline assembly failed to compile due to wrong syntax (at least with GCC 13).

Signed-off-by: Philipp Remy <philipp.remy@law-school.de>