PATCH: Dependencies: Add rpath handling to CMAKE_CORE_BUILD_FLAGS
If the user explicitly disables rpath handling when building AliceVision by setting ALICEVISION_USE_RPATH=OFF, respect this in the embedded dependencies.

This mainly fixes some older dependencies, where this is not set to ON by default and therefore results in a library with absolute install names (which drastically reduces portability of the resulting binary).

Signed-off-by: Philipp Remy <philipp.remy@law-school.de>