PATCH: Dependencies: Do not build tests for Ceres
As for other dependencies, do not build the tests for Ceres (as they are not needed and not tested). Besides from saving compilation time, this allows for cross-compiling Ceres on macOS. Otherwise, Ceres will build an embedded glog but does not pass through CMAKE_OSX_ARCHITECTURES, causing linking errors because of mismatched architectures.

Signed-off-by: Philipp Remy <philipp.remy@law-school.de>