PATCH: Dependencies: Build an embedded pybind11 if OpenImageIO is build
As we currently set -DUSE_PYTHON=ON when building OpenImageIO, it has a hard dependency on pybind11. Building this currently defaults to OFF. If no external pybind11 is found, this will cause a CMake error. As the project should be able to be build without any external dependencies in its standard configuration, build an embedded pybind11 by default.

Signed-off-by: Philipp Remy <philipp.remy@law-school.de>