PATCH: Add support for CMAKE_OSX_ARCHITECTURES
This commit introduces logic to ensure that CMAKE_OSX_ARCHITECTURES is always set on Apple targets. It can either be specified on the CLI (i.e., in case of cross-compilation) or left blank (in which case it is set to CMAKE_HOST_SYSTEM_PROCESSOR).

This will be used throughout the project to determine for which architecture to compile on Apple targets.

Signed-off-by: Philipp Remy <philipp.remy@law-school.de>