FIX: Dependencies: Explicitly use "-headerpad_max_install_names" on Apple targets where necessary
As it might be desirable to later change the embedded rpaths of a binary, "install_name_tool" expects enough padding between the header and the first data section of a Mach-O file. This caused issues when changing the rpaths of the following dependencies, which added little to no padding to the Mach-O header:

- libtiff
- ffmpeg

Pass "-headerpad_max_install_names" explicitly as linker flags to these dependencies when building them on Apple platforms. This adds the maximum amount of padding to the header (per ld man page: "Automatically adds space for future expansion of load commands such that all paths could expand to MAXPATHLEN.").

Signed-off-by: Philipp Remy <philipp.remy@law-school.de>